requirements.txt
fastapi
uvicorn
groq
python-dotenv
beautifulsoup4
requests
PyPDF2
python-docx